fix: rename to User Details
